+++ codegen/src/main/java/io/improbable/keanu/codegen/python/datamodel/VertexConstructor.java
@@ -1,0 +1,15 @@
+package io.improbable.keanu.codegen.python.datamodel;
+
+import lombok.Value;
+
+@Value
+public class VertexConstructor {
+
+    private String javaClass;
+    private String pythonClass;
+    private String pythonVertexClass;
+    private String pythonTypedParameters;
+    private String pythonParameters;
+    private String docString;
+
+}