+++ keanu-project/src/main/java/io/improbable/keanu/algorithms/mcmc/proposal/PriorProposalDistribution.java
@@ -1,5 +1,6 @@
 package io.improbable.keanu.algorithms.mcmc.proposal;
 
+import com.google.common.base.Preconditions;
 import io.improbable.keanu.KeanuRandom;
 import io.improbable.keanu.algorithms.Variable;
 import io.improbable.keanu.algorithms.VariableReference;
@@ -22,7 +23,8 @@ public class PriorProposalDistribution implements ProposalDistribution {
     }
 
     public PriorProposalDistribution(Collection<Vertex> vertices, List<ProposalListener> listeners) {
-        vertexLookup = vertices.stream().collect(Collectors.toMap(v -> v.getReference(), v -> v));
+        checkAllVerticesAreProbabilistic(vertices);
+        vertexLookup = vertices.stream().collect(Collectors.toMap(Vertex::getReference, v -> v));
         this.proposalNotifier = new ProposalNotifier(listeners);
 
     }
@@ -44,11 +46,17 @@ public class PriorProposalDistribution implements ProposalDistribution {
 
     private <T> void setFor(Variable<T, ?> variable, KeanuRandom random, Proposal proposal) {
         Vertex<T> vertex = vertexLookup.get(variable.getReference());
-        proposal.setProposal(variable, vertex.sample(random));
+        proposal.setProposal(variable, (((Probabilistic<T>) vertex).sample(random)));
     }
 
     @Override
     public void onProposalRejected() {
         proposalNotifier.notifyProposalRejected();
+    }
+
+    private void checkAllVerticesAreProbabilistic(Collection<Vertex> vertices) {
+        for (Vertex v : vertices) {
+            Preconditions.checkArgument(v instanceof Probabilistic, "Prior proposal vertices must be probabilistic. Vertex is: " + v);
+        }
     }
 }